requirements.txt
fastapi
uvicorn
cassandra-driver
python-dotenv
requests
pytest
email-validator